+++ CHANGES.md
@@ -32,6 +32,7 @@ Changes in 5.8
 - Hide Sort Order, Comments, Grade Posting Dates (if MP not Graded) from non editing users in MarkingPeriods.php
 - Moodle drop student use `enrol_manual_unenrol_users` WS function in Moodle/functions.php, MassDrops.php, Schedule.php
 - Display note messages if any in Student.php & User.php
+- Fix SQL error for numeric Student Fields of Number type in AssignOtherInfo.php
 
 Changes in 5.7.6
 ----------------